@@ -1,25 +1,22 @@
-using System;
 using System.ComponentModel.DataAnnotations;
 using System.Security.Claims;
-using UserManagementUiDemo.Models.Entities;
 
-namespace UserManagementUiDemo.Models.InputModels
+namespace UserManagementUiDemo.Models.InputModels;
+
+public class UserEditClaimInputModel
 {
-    public class UserEditClaimInputModel
-    {
 
-        [Required(ErrorMessage = "Il claim type è obbligatorio"),
-        MaxLength(255, ErrorMessage = "Il claim type è troppo lungo"),
-        Display(Name = "Claim Type")]
-        public string Type { get; set; }
+	[Required(ErrorMessage = "Il claim type è obbligatorio"),
+	MaxLength(255, ErrorMessage = "Il claim type è troppo lungo"),
+	Display(Name = "Claim Type")]
+	public string Type { get; set; }
 
-        [MaxLength(255, ErrorMessage = "Il claim value è troppo lungo"),
-        Display(Name = "Claim Value")]
-        public string Value { get; set; }
+	[MaxLength(255, ErrorMessage = "Il claim value è troppo lungo"),
+	Display(Name = "Claim Value")]
+	public string Value { get; set; }
 
-        public Claim ToClaim()
-        {
-            return new(Type, Value ?? string.Empty);
-        }
-    }
+	public Claim ToClaim()
+	{
+		return new(Type, Value ?? string.Empty);
+	}
 }
