@@ -1,80 +1,62 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.EntityFrameworkCore;
-using UserManagementUiDemo.Data;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using System;
+using UserManagementUiDemo.Data;
 using UserManagementUiDemo.Models.Entities;
 
-namespace UserManagementUiDemo
+namespace UserManagementUiDemo;
+
+public class Startup(IConfiguration configuration)
 {
-    public class Startup
-    {
-        public Startup(IConfiguration configuration)
-        {
-            Configuration = configuration;
-        }
+	public IConfiguration Configuration { get; } = configuration;
 
-        public IConfiguration Configuration { get; }
+	public void ConfigureServices(IServiceCollection services)
+	{
+		services.AddDbContext<ApplicationDbContext>(options
+			=> options.UseSqlite(Configuration.GetConnectionString("DefaultConnection")));
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddDbContext<ApplicationDbContext>(options =>
-                options.UseSqlite(
-                    Configuration.GetConnectionString("DefaultConnection")));
-            services.AddDatabaseDeveloperPageExceptionFilter();
+		services.AddDatabaseDeveloperPageExceptionFilter();
+		services.AddDefaultIdentity<ApplicationUser>(options =>
+		{
+			options.SignIn.RequireConfirmedAccount = true;
+			options.Lockout.MaxFailedAccessAttempts = 5;
+			options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(5);
+			options.Lockout.AllowedForNewUsers = true;
+		}).AddEntityFrameworkStores<ApplicationDbContext>();
 
-            services
-                    .AddDefaultIdentity<ApplicationUser>(options => 
-                    {
-                        options.SignIn.RequireConfirmedAccount = true;
-                        options.Lockout.MaxFailedAccessAttempts = 5;
-                        options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(5);
-                        options.Lockout.AllowedForNewUsers = true;
-                    })
-                    .AddEntityFrameworkStores<ApplicationDbContext>();
-            services.AddControllersWithViews();
-            services.AddRazorPages();
-        }
+		services.AddControllersWithViews();
+		services.AddRazorPages();
+	}
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseMigrationsEndPoint();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
-                app.UseHsts();
-            }
-            app.UseHttpsRedirection();
-            app.UseStaticFiles();
+	public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+	{
+		if (env.IsDevelopment())
+		{
+			app.UseDeveloperExceptionPage();
+			app.UseMigrationsEndPoint();
+		}
+		else
+		{
+			app.UseExceptionHandler("/Home/Error");
+			app.UseHsts();
+		}
 
-            app.UseRouting();
+		app.UseHttpsRedirection();
+		app.UseStaticFiles();
 
-            app.UseAuthentication();
-            app.UseAuthorization();
+		app.UseRouting();
 
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute(
-                    name: "default",
-                    pattern: "{controller=Home}/{action=Index}/{id?}");
-                endpoints.MapRazorPages();
-            });
-        }
-    }
+		app.UseAuthentication();
+		app.UseAuthorization();
+
+		app.UseEndpoints(endpoints =>
+		{
+			endpoints.MapControllerRoute(name: "default", pattern: "{controller=Home}/{action=Index}/{id?}");
+			endpoints.MapRazorPages();
+		});
+	}
 }
